+++ src/main/java/com/aaronpmaus/jProt/metrics/Metrics.java
@@ -126,8 +126,10 @@ public class Metrics{
     this.alphaStrucID = prot1.getProteinName();
     this.betaStrucID = prot2.getProteinName();
     // build the distance matrices out of the residues that were aligned
-    this.alphaDistancesMatrix = prot1.calculateCarbonAlphaDistanceMatrix(prot1Mask);
-    this.betaDistancesMatrix = prot2.calculateCarbonAlphaDistanceMatrix(prot2Mask);
+    //this.alphaDistancesMatrix = prot1.calculateCarbonAlphaDistanceMatrix(prot1Mask);
+    //this.betaDistancesMatrix = prot2.calculateCarbonAlphaDistanceMatrix(prot2Mask);
+    this.alphaDistancesMatrix = DistanceMatrixCalculator.calculateDistanceMatrix(prot1, prot1Mask);
+    this.betaDistancesMatrix = DistanceMatrixCalculator.calculateDistanceMatrix(prot2, prot2Mask);
     calculateDifferencesMatrix();
   }